- worked on the product list item layout - some refactorings
