+++ app/src/main/java/com/spaetimc/presentation/scan/productlist/ProductListListener.kt
@@ -1,4 +1,4 @@
-package com.spaetimc.presentation.scan
+package com.spaetimc.presentation.scan.productlist
 
 interface ProductListListener {
     fun doStuff()
