+++ app/src/main/res/drawable/product_placeholder.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="411.287dp"
+    android:height="411.287dp"
+    android:viewportWidth="411.287"
+    android:viewportHeight="411.287">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M230.789,102.56c16.934,4.068 39.322,6.31 63.04,6.31s46.105,-2.241 63.04,-6.31c6.036,-1.451 11.188,-3.085 15.469,-4.913v15.846c-1.459,1.568 -7.658,5.392 -23.516,8.693c-15.186,3.163 -34.716,4.904 -54.992,4.904c-26.479,0 -51.553,-3.002 -67.071,-8.03c-4.2,-1.361 -8.714,0.941 -10.076,5.145s0.941,8.714 5.145,10.076c17.278,5.599 43.522,8.81 72.003,8.81c23.717,0 46.104,-2.241 63.039,-6.311c2.116,-0.509 4.123,-1.04 6.022,-1.594l-9.659,58.25c-0.723,4.358 2.226,8.479 6.584,9.201c0.443,0.073 0.884,0.108 1.318,0.108c3.843,0 7.233,-2.776 7.883,-6.692l11.365,-68.542c5.328,-3.754 7.955,-8.139 7.955,-13.208v-34.22c0,-10.036 -10.294,-17.389 -31.469,-22.477c-16.934,-4.069 -39.322,-6.311 -63.04,-6.311s-46.106,2.241 -63.04,6.311c-21.176,5.088 -31.47,12.44 -31.47,22.477S209.613,97.471 230.789,102.56zM238.835,72.199c15.187,-3.162 34.717,-4.903 54.994,-4.903s39.808,1.741 54.994,4.903c13.571,2.826 20.067,6.034 22.573,7.884c-2.506,1.85 -9.003,5.057 -22.574,7.883c-15.187,3.162 -34.717,4.903 -54.993,4.903s-39.808,-1.741 -54.994,-4.903c-13.571,-2.826 -20.068,-6.034 -22.574,-7.883C218.766,78.233 225.263,75.025 238.835,72.199zM353.888,217.121c-0.585,-0.22 -1.08,-0.414 -1.503,-0.579c-2.726,-1.066 -8.402,-3.287 -12.997,3.092c-0.321,0.447 -0.625,0.869 -1.193,1.476c-3.699,3.936 -15.286,7.92 -30.273,7.92c-19.048,0 -29.67,-6.101 -30.852,-8.642c0.252,-0.455 1.17,-1.563 3.384,-2.875c3.018,-1.787 4.52,-5.332 3.703,-8.743c-0.815,-3.412 -3.758,-5.894 -7.258,-6.123c-9.094,-0.594 -18.457,-0.896 -27.831,-0.896c-70.096,0 -141.055,17.042 -141.055,49.615v110.305c0,32.573 70.959,49.616 141.055,49.616s141.056,-17.043 141.056,-49.616V251.366C390.124,233.468 367.422,222.229 353.888,217.121zM374.124,361.671c0,5.239 -8.736,14.091 -33.359,21.792c-24.379,7.625 -56.943,11.824 -91.696,11.824c-34.752,0 -67.317,-4.199 -91.695,-11.824c-24.623,-7.701 -33.359,-16.553 -33.359,-21.792v-86.022c24.881,16.706 75.188,25.333 125.055,25.333c49.868,0 100.175,-8.627 125.056,-25.334V361.671zM160.173,274.008c3.613,-2.386 12.128,-5.547 24.899,-5.547c15.318,0 24.512,4.547 26.472,6.851c-1.318,1.494 -6.263,4.615 -16.03,6.103C182.624,279.596 170.691,277.107 160.173,274.008zM364.301,262.953c-0.262,-4.182 -3.736,-7.492 -7.984,-7.492h-2.012c-4.418,0 -8,3.582 -8,8c0,2.727 1.363,5.134 3.446,6.578c-2.727,1.053 -5.717,2.096 -8.987,3.118c-24.378,7.625 -56.943,11.824 -91.696,11.824c-8.234,0 -16.347,-0.235 -24.255,-0.697c1.939,-2.711 2.968,-5.727 2.968,-8.962c0,-15.009 -21.484,-22.861 -42.708,-22.861c-17.617,0 -35.414,5.411 -40.954,15.834c-14.597,-6.331 -20.105,-12.769 -20.105,-16.929c0,-3.802 4.601,-9.506 16.535,-15.292c0.444,3.996 3.835,7.104 7.95,7.104h2.012c4.418,0 8,-3.582 8,-8c0,-2.044 -0.767,-3.908 -2.027,-5.322c0.294,-0.095 0.591,-0.188 0.89,-0.281c24.379,-7.625 56.943,-11.824 91.695,-11.824c4.064,0 8.124,0.059 12.159,0.176c-0.152,0.917 -0.205,1.744 -0.205,2.449c0,16.186 23.594,24.653 46.899,24.653c17.985,0 33.777,-4.771 41.477,-12.493c14.397,5.634 24.726,13.464 24.726,18.83C374.124,254.351 371.29,258.506 364.301,262.953zM233.733,233.703c0,4.418 -3.582,8 -8,8h-2.014c-4.418,0 -8,-3.582 -8,-8s3.582,-8 8,-8h2.014C230.151,225.703 233.733,229.285 233.733,233.703zM278.676,265.674c0,4.418 -3.582,8 -8,8h-2.013c-4.418,0 -8,-3.582 -8,-8s3.582,-8 8,-8h2.013C275.094,257.674 278.676,261.256 278.676,265.674zM151.834,185.931v-33.247c0,-13.372 -1.772,-23.268 -6.118,-34.157c-1.637,-4.104 0.362,-8.758 4.466,-10.396c4.104,-1.639 8.757,0.362 10.396,4.466c5.087,12.748 7.257,24.736 7.257,40.087v33.247c0,4.418 -3.582,8 -8,8S151.834,190.349 151.834,185.931zM86.642,369.803c-18.015,0 -32.67,-13.827 -32.67,-30.822V149.862c0,-23.226 5.65,-34.455 15.013,-53.061c0.515,-1.023 1.043,-2.074 1.584,-3.153h97.585c0.541,1.08 1.069,2.131 1.585,3.155c9.36,18.604 15.011,29.834 15.011,53.059v28.545c0,4.418 3.582,8 8,8s8,-3.582 8,-8v-28.545c0,-27.023 -7.013,-40.96 -16.718,-60.25c-1.19,-2.366 -2.42,-4.809 -3.686,-7.371c-0.041,-0.087 -0.084,-0.174 -0.128,-0.259c-3.739,-7.577 -7.798,-16.205 -12.13,-26.976l-0.001,-2.356c3.566,-3.958 5.739,-9.195 5.739,-14.93V22.326C173.827,10.016 163.811,0 151.501,0H87.219C74.909,0 64.894,10.016 64.894,22.326v12.54l-21.061,-8.665c-4.083,-1.679 -8.761,0.268 -10.442,4.354c-1.681,4.086 0.269,8.761 4.354,10.442l23.399,9.627L24.941,73.134c-3.753,2.333 -4.903,7.266 -2.57,11.018c2.332,3.752 7.266,4.904 11.018,2.57L65.73,66.614c-2.523,5.688 -4.937,10.729 -7.229,15.376c-0.041,0.08 -0.081,0.161 -0.12,0.243c-1.267,2.563 -2.497,5.009 -3.688,7.377c-9.707,19.29 -16.72,33.229 -16.72,60.252V338.98c0,25.818 21.833,46.822 48.67,46.822c4.418,0 8,-3.582 8,-8S91.06,369.803 86.642,369.803zM80.894,22.326c0,-3.488 2.838,-6.326 6.325,-6.326h64.281c3.488,0 6.326,2.838 6.326,6.326v15.395c0,1.957 -0.894,3.71 -2.294,4.872c-0.167,0.116 -0.33,0.238 -0.488,0.366c-1.012,0.688 -2.231,1.089 -3.544,1.089H87.219c-1.301,0 -2.511,-0.395 -3.517,-1.07c-0.173,-0.142 -0.351,-0.274 -0.533,-0.4c-1.39,-1.161 -2.275,-2.907 -2.275,-4.856V22.326zM85.869,60.008c0.447,0.026 0.896,0.04 1.351,0.04h64.281c0.454,0 0.906,-0.014 1.354,-0.041c2.606,6.529 5.148,12.369 7.569,17.642H78.298C80.72,72.376 83.262,66.537 85.869,60.008z"/>
+</vector>
