+++ app/src/main/res/layout/activity_scan.xml
@@ -15,7 +15,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        tools:context=".presentation.search.MainActivity"
-        tools:listitem="@layout/searchresult_entry" />
+        tools:context=".presentation.scan.ScanActivity"
+        tools:listitem="@layout/product_list_entry" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
